+++ contracts/finality/src/state/mod.rs
@@ -3,6 +3,7 @@ pub mod config;
 pub mod finality;
 pub mod pruning;
 pub mod public_randomness;
+pub mod rate_limiting;
 
 /// `Bytes` is a vector of bytes
 pub type Bytes = Vec<u8>;